Downgrade julia to avoid m1 instabilities

+++ src/Models/FullyConnected.jl
@@ -50,7 +50,7 @@ function fully_connected_neural_network(
 end
 
 function model_input(::Type{FullyConnected}, data::Vector{Dict{String,Any}})
-    return Float32.(hcat(map(d -> [d["parameters"][pd.key]..., d["parameters"][qd.key]...], data)...))
+    return Float64.(hcat(map(d -> [d["parameters"][pd.key]..., d["parameters"][qd.key]...], data)...))
 end
 
 function model_factory(::Type{FullyConnected}, size_in::Union{Int64,Tuple}, size_out::Int, num_layers::Int; kwargs...)
